% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gn.R
\name{gn_geom_defaults}
\alias{gn_geom_defaults}
\title{Change geom defaults from black to custom lights for the GreyNoise theme}
\usage{
gn_geom_defaults()
}
\description{
Change geom defaults from black to custom lights for the GreyNoise theme
}
