% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gn-plot.R
\name{gn_port_summary}
\alias{gn_port_summary}
\title{Generate a simple port summary plot}
\usage{
gn_port_summary(port, protocol = c("tcp", "udp"), save = NULL,
  gn_key = Sys.getenv("GREYNOISE_API_KEY"))
}
\arguments{
\item{port, protocol, gn_key}{same as \code{\link[=research_time_series]{research_time_series()}}}

\item{save}{if not \code{NULL}, the full path to save the file (with \code{.png} extension)}
}
\description{
Generate a simple port summary plot
}
