% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-ips.r
\name{gn_query_by_ip}
\alias{gn_query_by_ip}
\title{Query all tags associated with a given IP address}
\usage{
gn_query_by_ip(ip)
}
\description{
GreyNoise adds scanner tags to IP addresses. This function retrieves
all tags currently in use.
}
