% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-tags.R
\name{gn_query_by_tag}
\alias{gn_query_by_tag}
\title{Query all IPs that have a given tag}
\usage{
gn_query_by_tag(tag)
}
\description{
GreyNoise adds scanner tags to IP addresses. This function retrieves
all tags currently in use.
}
