% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gn.R
\docType{data}
\name{gn_text_col}
\alias{gn_text_col}
\title{The GreyNoise default font color}
\format{An object of class \code{character} of length 1.}
\usage{
gn_text_col
}
\description{
The GreyNoise default font color
}
\keyword{datasets}
