% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greynoise-package.R
\docType{package}
\name{greynoise}
\alias{greynoise}
\title{Query 'GreyNoise Intelligence 'API'}
\description{
Tools are provided to query the 'GreyNoise Intelligence 'API'.
}
\author{
Bob Rudis (bob@rud.is)
}
