% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-helpers.R
\name{install_cmdline_helpers}
\alias{install_cmdline_helpers}
\title{Install command-line helpers}
\usage{
install_cmdline_helpers(prefix = "/usr/local/bin")
}
\arguments{
\item{prefix}{prefix to \code{bin} directory. Defaults to \verb{/usr/local/bin}}
}
\description{
Install command-line helpers
}
\note{
You need the proper permissions for this to work and it assumes a linux environment. You
}
