% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/research-actors.R
\name{research_actors}
\alias{research_actors}
\title{Retrieve research benign actors}
\usage{
research_actors(gn_key = Sys.getenv("GREYNOISE_API_KEY"))
}
\arguments{
\item{gn_key}{your greynoise API key. This function uses the authenticated V2
API endpoint and will look for the API key in the \code{GREYNOISE_API_KEY}
environment variable.}
}
\description{
Retrieve research benign actors
}
