% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsrch-stats.R
\name{research_stats}
\alias{research_stats}
\title{"Research API" bare-bones "stats" endpoint}
\usage{
research_stats(path, gn_key = Sys.getenv("GREYNOISE_API_KEY"))
}
\arguments{
\item{path}{path (e.g. \code{top/http/path})}

\item{gn_key}{your greynoise API key. This function uses the authenticated V2
API endpoint and will look for the API key in the \code{GREYNOISE_API_KEY}
environment variable.}
}
\description{
"Research API" bare-bones "stats" endpoint
}
