% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsrch-scan-ts.R
\name{research_time_series}
\alias{research_time_series}
\title{Retrieve (optionally summarised) scan history for a given port/protocol}
\usage{
research_time_series(port, protocol = c("tcp", "udp"), summarise = TRUE,
  gn_key = Sys.getenv("GREYNOISE_API_KEY"))
}
\arguments{
\item{port}{port to query for}

\item{protocol}{protocol (defaults to `tcp``)}

\item{summarise}{logical. If \code{TRUE} the \code{time_series} endpoint will be used.
If \code{FALSE} the \code{raw} endpoint will be used.}

\item{gn_key}{your greynoise API key. This function uses the authenticated V2
API endpoint and will look for the API key in the \code{GREYNOISE_API_KEY}
environment variable.}
}
\description{
Retrieve (optionally summarised) scan history for a given port/protocol
}
