% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gn.R
\name{theme_gn}
\alias{theme_gn}
\title{GreyNoise ggplot2 theme}
\usage{
theme_gn(
  base_family = "Lato",
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_margin = 10,
  subtitle_family = "Lato",
  subtitle_size = 13,
  subtitle_face = "plain",
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = "Lato",
  caption_size = 9,
  caption_face = "plain",
  caption_margin = 10,
  axis_text_family = "IBMPlexSans",
  axis_text_size = base_size,
  axis_title_family = base_family,
  axis_title_size = 9,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30),
  grid = TRUE,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{base font family and size}

\item{base_size}{base font family and size}

\item{plot_title_family}{plot tilte family, face, size and margin}

\item{plot_title_size}{plot tilte family, face, size and margin}

\item{plot_title_face}{plot tilte family, face, size and margin}

\item{plot_title_margin}{plot tilte family, face, size and margin}

\item{subtitle_family}{plot subtitle family, face and size}

\item{subtitle_size}{plot subtitle family, face and size}

\item{subtitle_face}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family}{facet label font family, face and size}

\item{strip_text_size}{facet label font family, face and size}

\item{strip_text_face}{facet label font family, face and size}

\item{caption_family}{plot caption family, face, size and margin}

\item{caption_size}{plot caption family, face, size and margin}

\item{caption_face}{plot caption family, face, size and margin}

\item{caption_margin}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family}{axis title font family, face and size}

\item{axis_title_size}{axis title font family, face and size}

\item{axis_title_face}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of \verb{[blmcrt]}}

\item{plot_margin}{plot margin (specify with \link[ggplot2:margin]{ggplot2::margin})}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{ticks if \code{TRUE} add ticks}
}
\description{
GreyNoise ggplot2 theme
}
