% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag-list.r
\name{gn_list_tags}
\alias{gn_list_tags}
\title{List GreyNoise Intelligence Tags}
\usage{
gn_list_tags()
}
\description{
GreyNoise adds scanner tags to IP addresses. This function retrieves
all tags currently in use.
}
